requirements.txt
flask
flickrapi
requests
pillow
gunicorn